chore: split browser telemetry into cmd/browsers_telemetry.go
